clean warnings

+++ src/context/mod.rs
@@ -192,7 +192,7 @@ impl Context {
                 raw::RedisModule_ReplyWithStringBuffer.unwrap()(
                     self.ctx,
                     s.as_ptr().cast::<c_char>(),
-                    s.len() as usize,
+                    s.len(),
                 )
                 .into()
             },
@@ -205,7 +205,7 @@ impl Context {
                 raw::RedisModule_ReplyWithStringBuffer.unwrap()(
                     self.ctx,
                     s.as_ptr().cast::<c_char>(),
-                    s.len() as usize,
+                    s.len(),
                 )
                 .into()
             },
